code/Ingestion/services/storage/v1.0.0/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pymilvus==2.4.1
python-dotenv==1.0.1
marshmallow<4.0.0  # Fix for environs compatibility
